services/nl-sql-service/pyproject.toml
# services/nl-sql-service/pyproject.toml
[project]
name = "nl-sql-service"
version = "0.1.0"
description = "Natural Language to SQL query service for Rice Market ERP"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "psycopg2-binary==2.9.9",
    "sqlalchemy==2.0.23",
    "langchain==0.2.16",
    "langchain-community==0.2.16",
    "langchain-google-vertexai==1.0.6",
    "google-cloud-aiplatform==1.38.0",
    "python-dotenv==1.0.0",
    "httpx==0.25.2",
    "sqlparse==0.4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
]
